refactor: always tag top level const value
