+++ lib/WebpackOptionsApply.js
@@ -59,10 +59,14 @@ const URLPlugin = require("./dependencies/URLPlugin");
 
 const WorkerPlugin = require("./dependencies/WorkerPlugin");
 
+const ConstValueParserPlugin = require("./javascript/ConstValueParserPlugin");
+
 const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
 const JavascriptParser = require("./javascript/JavascriptParser");
+const JavascriptParserApplyPlugin = require("./javascript/JavascriptParserApplyPlugin");
 
 const JsonModulesPlugin = require("./json/JsonModulesPlugin");
+const CircularModulesPlugin = require("./optimize/CircularModulesPlugin");
 
 const ChunkPrefetchPreloadPlugin = require("./prefetch/ChunkPrefetchPreloadPlugin");
 
@@ -391,6 +395,13 @@ class WebpackOptionsApply extends OptionsApply {
 		}
 
 		new JavascriptModulesPlugin().apply(compiler);
+		new JavascriptParserApplyPlugin(
+			new ConstValueParserPlugin({
+				inlineExports: Boolean(options.optimization.inlineExports)
+			})
+		).apply(compiler);
+		new CircularModulesPlugin().apply(compiler);
+
 		new JsonModulesPlugin().apply(compiler);
 		new AssetModulesPlugin({
 			sideEffectFree: options.experiments.futureDefaults
@@ -639,10 +650,8 @@ class WebpackOptionsApply extends OptionsApply {
 		}
 		if (options.optimization.inlineExports) {
 			const InlineExportsPlugin = require("./optimize/InlineExportsPlugin");
-			const CircularModulesPlugin = require("./optimize/CircularModulesPlugin");
 
 			new InlineExportsPlugin().apply(compiler);
-			new CircularModulesPlugin().apply(compiler);
 		}
 		if (options.optimization.mangleExports) {
 			const MangleExportsPlugin = require("./optimize/MangleExportsPlugin");
