+++ lib/optimize/InlineExports.js
@@ -16,9 +16,6 @@ const { propertyAccess } = require("../util/property");
 
 const MAX_INLINE_BYTES = 6;
 
-// Parser tag for top-level `const X = <primitive>` bindings
-const CONST_BINDING_TAG = Symbol("const binding");
-
 /** @typedef {"null" | "undefined" | "boolean" | "number" | "string"} InlinedValueKind */
 
 class InlinedValue {
@@ -167,7 +164,6 @@ const isExportInlined = (moduleGraph, module, ids, runtime) => {
 	return moduleGraph.getExportsInfo(module).hasInlinedUsedName(ids, runtime);
 };
 
-module.exports.CONST_BINDING_TAG = CONST_BINDING_TAG;
 module.exports.InlinedUsedName = InlinedUsedName;
 module.exports.InlinedValue = InlinedValue;
 module.exports.MAX_INLINE_BYTES = MAX_INLINE_BYTES;
