+++ lib/optimize/InlineExportsPlugin.js
@@ -10,12 +10,15 @@ const {
 	JAVASCRIPT_MODULE_TYPE_AUTO,
 	JAVASCRIPT_MODULE_TYPE_ESM
 } = require("../ModuleTypeConstants");
-const ConstValueParserPlugin = require("./ConstValueParserPlugin");
+const BasicEvaluatedExpression = require("../javascript/BasicEvaluatedExpression");
+const { CONST_BINDING_TAG } = require("../javascript/ConstValueParserPlugin");
 const { InlinedUsedName, enableInlineExports } = require("./InlineExports");
 
+/** @typedef {import("./InlineExports").InlinedValue} InlinedValue */
 /** @typedef {import("../Compiler")} Compiler */
 /** @typedef {import("../ExportsInfo")} ExportsInfo */
 /** @typedef {import("../javascript/JavascriptParser")} JavascriptParser */
+/** @typedef {import("../javascript/JavascriptModule").JavascriptModuleBuildInfo} JavascriptModuleBuildInfo */
 
 const PLUGIN_NAME = "InlineExportsPlugin";
 
@@ -36,7 +39,38 @@ class InlineExportsPlugin {
 				 * @returns {void}
 				 */
 				const handler = (parser) => {
-					new ConstValueParserPlugin().apply(parser);
+					parser.hooks.program.tap(PLUGIN_NAME, () => {
+						const buildInfo =
+							/** @type {JavascriptModuleBuildInfo | undefined} */
+							(parser.state.module.buildInfo);
+						if (buildInfo) buildInfo.inlineExports = true;
+					});
+
+					// Propagate inlined constant through evaluator so chained constants and uses see the literal
+					parser.hooks.evaluateIdentifier
+						.for(CONST_BINDING_TAG)
+						.tap(PLUGIN_NAME, (expr) => {
+							const tagData =
+								/** @type {{ value?: InlinedValue } | undefined} */
+								(parser.currentTagData);
+							if (!tagData || !tagData.value) return;
+							const { value } = tagData;
+							const eval_ = new BasicEvaluatedExpression().setRange(
+								/** @type {[number, number]} */ (expr.range)
+							);
+							switch (value.kind) {
+								case "null":
+									return eval_.setNull();
+								case "undefined":
+									return eval_.setUndefined();
+								case "boolean":
+									return eval_.setBoolean(/** @type {boolean} */ (value.value));
+								case "number":
+									return eval_.setNumber(/** @type {number} */ (value.value));
+								case "string":
+									return eval_.setString(/** @type {string} */ (value.value));
+							}
+						});
 				};
 				normalModuleFactory.hooks.parser
 					.for(JAVASCRIPT_MODULE_TYPE_AUTO)
